A2A Agentic Communication/requirements.txt
a2a-sdk[http-server]>=0.3.25
uvicorn>=0.30.0
httpx>=0.27.0
